+++ Sistema-Gestor-Obras-PagesFamilia/Sistema-Gestor-Obras-PagesFamilia/Repositories/IOficioRepository.cs
@@ -1,0 +1,5 @@
+using Sistema_Gestor_Obras_PagesFamilia.Models;
+
+namespace Sistema_Gestor_Obras_PagesFamilia.Repositories;
+
+public interface IOficioRepository : IRepository<Oficio> { }